+++ .gitignore
@@ -1,4 +1,4 @@
-test/dbs/full
+benchmark/*.mmdb
 testing.js
 node_modules
 .DS_Store
