+++ lib/reader.js
@@ -1,4 +1,5 @@
 var fs = require('fs');
+var LRU = require('lru-cache');
 var Metadata = require('./metadata');
 var Decoder = require('./decoder');
 var ipUtil = require('./ip');
@@ -25,6 +26,12 @@ function Reader() {}
 
 Reader.open = function(database) {
   var reader = new Reader();
+
+  reader.cache = LRU({
+    max: 50000,
+    maxAge: 60 * 60 * 1000, // 1hr
+  });
+
   var start, metadataDecoder, metadataArray;
 
   reader.fileHandle = fs.readFileSync(database);
@@ -173,7 +180,13 @@ Reader.prototype.resolveDataPointer = function(pointer) {
   //     $offset_in_file = ( $record_value - $node_count )
   //                       + $search_tree_size_in_bytes
 
+  if (this.cache.has(pointer)) {
+    return this.cache.get(pointer);
+  }
+
   var resolved = pointer - this.metadata.nodeCount + this.metadata.searchTreeSize;
 
-  return this.decoder.decode(resolved)[0];
+  var result = this.decoder.decode(resolved)[0];
+  this.cache.set(pointer, result);
+  return result;
 };
